Add styles to the navbar
