+++ iot/angular/iot/src/app/app.component.html
@@ -1,2 +1,2 @@
-<app-navigation></app-navigation>
+<app-navbar class="mat-elevation-z6"></app-navbar>
 <router-outlet></router-outlet>
