+++ iot/angular/iot/src/app/app.module.ts
@@ -4,14 +4,13 @@ import { FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
 
 import { AppComponent } from './app.component';
-import { NavigationComponent } from './shared/navigation/navigation.component';
+import { NavbarComponent } from './shared/navbar/navbar.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AppRoutingModule } from './/app-routing.module';
 
 import { MatListModule } from '@angular/material/list';
 import { MatButtonModule, MatMenuModule } from '@angular/material';
 import { MatIconModule } from '@angular/material/icon';
-import { MatToolbarModule } from '@angular/material/toolbar';
 
 import { HistoryComponent } from './pages/history/history.component';
 import { HomeComponent } from './pages/home/home.component';
@@ -21,11 +20,11 @@ import { StatusComponent } from './pages/status/status.component';
 @NgModule({
   declarations: [
     AppComponent,
-    NavigationComponent,
+    NavbarComponent,
     HistoryComponent,
     HomeComponent,
     NotfoundComponent,
-    StatusComponent,
+    StatusComponent
   ],
   imports: [
     BrowserModule,
@@ -37,10 +36,9 @@ import { StatusComponent } from './pages/status/status.component';
     MatListModule,
     MatButtonModule,
     MatMenuModule,
-    MatIconModule,
-    MatToolbarModule,
+    MatIconModule
   ],
   providers: [],
   bootstrap: [AppComponent]
 })
-export class AppModule { }
+export class AppModule {}
