+++ iot/angular/iot/src/app/shared/navbar/_navbar-theme.scss
@@ -1,0 +1,15 @@
+@mixin nav-bar-theme($theme) {
+  $primary: map-get($theme, primary);
+  $accent: map-get($theme, accent);
+  $warn: map-get($theme, warn);
+  $background: map-get($theme, background);
+  $foreground: map-get($theme, foreground);
+
+  app-navbar {
+    color: mat-color($primary, default-contrast);
+
+    .navbar, .navbar-header {
+      background: mat-color($primary);
+    }
+  }
+}
