+++ iot/angular/iot/src/app/shared/navbar/navbar.component.scss
@@ -54,7 +54,7 @@
     display: flex;
   }
 
-  .navbar--github-logo {
+  .navbar-github-logo {
     margin: 0;
   }
 }
