+++ iot/angular/iot/src/app/shared/navbar/navbar.component.ts
@@ -1,0 +1,19 @@
+import { Component } from '@angular/core';
+import { SECTIONS } from '../documentation-items';
+
+const SECTIONS_KEYS = Object.keys(SECTIONS);
+
+@Component({
+  selector: 'app-navbar',
+  templateUrl: './navbar.component.html',
+  styleUrls: ['./navbar.component.scss']
+})
+export class NavbarComponent {
+  get sections() {
+    return SECTIONS;
+  }
+
+  get sectionKeys() {
+    return SECTIONS_KEYS;
+  }
+}
